backend/requirements.txt
fastapi==0.100.0
uvicorn
uvicorn==0.22.0
pandas==2.0.3
SQLAlchemy==2.0.19
python-socketio==5.8.0
